// src/interfaces/login.interface.tsx
// Generated by https://quicktype.io

export interface ResultLogin {
    ok:        boolean;
    msg:       string;
    ciudadano: Ciudadano;
    token:     string;
}

export interface Ciudadano {
    id:       number;
    dni:      string;
    nombre:   string;
    apellido: string;
    password: string;
    estado:   number;
}

export interface LoginData {
    dni:string;
    password:string
}

export interface RegisterCiudadano{
    nombre:string,
    apellido:string,
    dni:string,
    password:string
}